# powerapps/connectors/MILSTRIP-Local-Dev-API.swagger.json
{
  "basePath": "/api/v1",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "HTTPValidationError": {
      "properties": {
        "detail": {
          "items": {
            "$ref": "#/definitions/ValidationError"
          },
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "IntakeAcknowledgement": {
      "properties": {
        "received_at": {
          "format": "date-time",
          "type": "string"
        },
        "records": {
          "type": "integer"
        },
        "rejected": {
          "type": "integer"
        },
        "request_id": {
          "type": "string"
        },
        "requires_review": {
          "type": "integer"
        },
        "status": {
          "enum": [
            "RECEIVED",
            "PROCESSING",
            "REQUIRES_REVIEW",
            "VALID",
            "REJECTED",
            "FAILED"
          ],
          "type": "string"
        }
      },
      "required": [
        "request_id",
        "status",
        "received_at",
        "records",
        "rejected",
        "requires_review"
      ],
      "type": "object"
    },
    "IntakeRequest": {
      "properties": {
        "source_id": {
          "type": "string",
          "x-nullable": true
        },
        "source_text": {
          "maxLength": 1000000,
          "minLength": 1,
          "type": "string"
        },
        "source_type": {
          "pattern": "^(PASTE|FILE|FRESHSERVICE)$",
          "type": "string"
        },
        "submitted_by": {
          "type": "string",
          "x-nullable": true
        }
      },
      "required": [
        "source_type",
        "source_text"
      ],
      "type": "object"
    },
...
    "RequestPage": {
      "properties": {
        "items": {
          "items": {
            "$ref": "#/definitions/RequestSummary"
          },
          "type": "array"
        },
        "next_cursor": {
          "type": "string",
          "x-nullable": true
        }
      },
      "required": [
        "items",
        "next_cursor"
      ],
      "type": "object"
    },
    "RequestSummary": {
      "properties": {
        "completed_at": {
          "format": "date-time",
          "type": "string",
          "x-nullable": true
        },
        "received_at": {
          "format": "date-time",
          "type": "string"
        },
        "request_id": {
          "type": "string"
        },
        "source_id": {
          "type": "string",
          "x-nullable": true
        },
        "source_sha256": {
          "type": "string"
        },
        "source_type": {
          "type": "string"
        },
        "status": {
          "enum": [
            "RECEIVED",
            "PROCESSING",
            "REQUIRES_REVIEW",
            "VALID",
            "REJECTED",
            "FAILED"
          ],
          "type": "string"
        },
        "submitted_by": {
          "type": "string",
          "x-nullable": true
        }
      },
      "required": [
        "request_id",
        "source_type",
        "source_id",
        "source_sha256",
        "status",
        "submitted_by",
        "received_at",
        "completed_at"
      ],
      "type": "object"
    },
...
    "ValidationError": {
      "properties": {
        "ctx": {
          "type": "object"
        },
        "input": {},
        "loc": {
          "items": {
            "type": "string"
          },
          "type": "array"
        },
        "msg": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "required": [
        "loc",
        "msg",
        "type"
      ],
      "type": "object"
    },
...
  },
  "host": "127.0.0.1:8000",
  "info": {
    "title": "MILSTRIP Local Dev API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/intake/requests": {
      "get": {
        "operationId": "ListIntakeRequests",
        "parameters": [
          {
            "default": 50,
            "in": "query",
            "maximum": 100,
            "minimum": 1,
            "name": "limit",
            "required": false,
            "type": "integer"
          },
          {
            "enum": [
              "RECEIVED",
              "PROCESSING",
              "REQUIRES_REVIEW",
              "VALID",
              "REJECTED",
              "FAILED"
            ],
            "in": "query",
            "name": "status",
            "required": false,
            "type": "string",
            "x-nullable": true
          },
          {
            "in": "query",
            "maxLength": 2048,
            "name": "cursor",
            "required": false,
            "type": "string",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/RequestPage"
            }
          },
          "401": {
            "description": "Missing or invalid API credentials",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "403": {
            "description": "Outside laptop boundary",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "409": {
            "description": "Review conflict; reload before a new command",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          },
          "503": {
            "description": "Local configuration or database unavailable",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "List Requests"
      },
      "post": {
        "operationId": "CreateIntakeRequest",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IntakeRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/IntakeAcknowledgement"
            }
          },
          "401": {
            "description": "Missing or invalid API credentials",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "403": {
            "description": "Outside laptop boundary",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "409": {
            "description": "Review conflict; reload before a new command",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/HTTPValidationError"
            }
          },
          "503": {
            "description": "Local configuration or database unavailable",
            "schema": {
              "properties": {
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Create Intake Request"
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "security": [
    {
      "basic-auth": []
    }
  ],
  "securityDefinitions": {
    "basic-auth": {
      "type": "basic"
    }
  },
  "swagger": "2.0"
}